Load encrypted org files.
If A user can store his configuration files in a public repository,
and that his configuration files contain some private information, he
may want to encrypt them. As Emacs will automaticly do what is needed
for decrypt gpg files, we just have to open them, and let the magic
be.