Remove broken links in starter-kit-misc.org
Remove broken links in the 'Color Themes' section in
starter-kit-misc.org. Provide new link for pre-built themes.

Additional information provided on where themes should and/or will be
installed as well as further instructions required to activate a desired
theme.

Correct the heading title of section handling backup files to be more
explicit.